Fix passed pawns preliminarily
The evaluation was very weird before this change as there was something wrong with my passed pawns implementation.

I now include passed pawns eval only in endgame, and also include enemy pawns as a penalty.

Note that I still expect problems to occur with this in the endgame, and I will look into it in detail later.